+++ seiscat/database/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf8 -*-
+# SPDX-License-Identifier: GPL-3.0-or-later
+"""
+Database functions for seiscat.
+
+:copyright:
+    2022-2025 Claudio Satriano <satriano@ipgp.fr>
+:license:
+    GNU General Public License v3.0 or later
+    (https://www.gnu.org/licenses/gpl-3.0-standalone.html)
+"""
+from .get_set import seiscat_get, seiscat_set  # noqa