+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Copyright (c) 2022-2025 Claudio Satriano <satriano@ipgp.fr>
 - New option for `seiscat initdb` and `seiscat updatedb`: `--fromfile` to
   initialize or update the database from a CSV file
 - New command `seiscat run` to run a user-defined command on each event
+- New commands `seiscat get` and `seiscat set` to get and set the value of
+  a specfic event attribute
 - New command `seiscat logo` to print the beautiful, ascii-art SeisCat logo
 
 ## v0.8 - 2024-10-28